Added serialdevice.on decorator.
Also moved as_coroutine to util.py
